# Optimize flattening in apache airflow workspace