… crates/uv-distribution/src/metadata/build_requires.rs